refactor: PostDetail 컴포넌트 분리 및 width 수정
